better es

+++ mix.exs
@@ -90,7 +90,7 @@ defmodule NxHailo.MixProject do
   defp aliases do
     [
       setup: ["deps.get"],
-      "compile.download_models": [&download_yolov8_model/1]
+      "compile.download_models": [&download_yolov8_model/1, &download_resnet_model/1]
     ]
   end
 
@@ -101,7 +101,7 @@ defmodule NxHailo.MixProject do
       "https://raw.githubusercontent.com/ultralytics/ultralytics/refs/heads/main/ultralytics/cfg/datasets/coco.yaml"
 
     model_hef_url =
-      "https://hailo-model-zoo.s3.eu-west-2.amazonaws.com/ModelZoo/Compiled/v2.15.0/hailo8l/yolov8m.hef"
+      "https://hailo-model-zoo.s3.eu-west-2.amazonaws.com/ModelZoo/Compiled/v2.17.0/hailo8/yolov8m.hef"
 
     priv = Path.join(__DIR__, "priv")
 
@@ -109,6 +109,23 @@ defmodule NxHailo.MixProject do
 
     download_dataset_to_json_file(dataset_yml, Path.join(priv, "yolov8m_classes.json"))
     download_model(model_hef_url, Path.join(priv, "yolov8m.hef"))
+  end
+
+  defp download_resnet_model(_args) do
+    {:ok, _} = Application.ensure_all_started([:req])
+
+    dataset_yml =
+      "https://s3.amazonaws.com/deep-learning-models/image-models/imagenet_class_index.json"
+
+    model_hef_url =
+      "https://hailo-model-zoo.s3.eu-west-2.amazonaws.com/ModelZoo/Compiled/v2.17.0/hailo8/resnet_v1_50.hef"
+
+    priv = Path.join(__DIR__, "priv")
+
+    File.mkdir_p!(priv)
+
+    download_json_dataset_to_json_file(dataset_yml, Path.join(priv, "imagenet_class_index.json"))
+    download_model(model_hef_url, Path.join(priv, "resnet_v1_50.hef"))
   end
 
   defp download_dataset_to_json_file(url, filename) do
@@ -123,6 +140,22 @@ defmodule NxHailo.MixProject do
         |> Map.get("names")
         |> Enum.sort_by(fn {index, _name} -> index end)
         |> Enum.map(fn {_index, name} -> name end)
+        |> Jason.encode!()
+
+      File.write!(filename, contents)
+    end
+  end
+
+  defp download_json_dataset_to_json_file(url, filename) do
+    if File.exists?(filename) do
+      :ok
+    else
+      %{body: yaml_contents} = Req.get!(url)
+
+      contents =
+        yaml_contents
+        |> Enum.sort_by(fn {index, _name} -> index end)
+        |> Enum.map(fn {_index, [_cat, name]} -> name end)
         |> Jason.encode!()
 
       File.write!(filename, contents)
